+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+-   Added: `--config-basedir` option to CLI.
 -   Added: `ignore: ["consecutive-duplicates-with-different-values"]` option to `declaration-block-no-duplicate-properties`.
 -   Added: `--report-needless-disables` and `reportNeedlessDisables` option.
 -   Added: `--ignore-disables` and `ignoreDisables` option.